Renaming CurveImpl_DNC to CurveBackend_DNC and moving to package de.uni_kl.cs.discodnc. Renaming method getCurve to getCurveFactory.
