+++ src/main/java/de/uni_kl/cs/discodnc/CurveBackend.java
@@ -36,7 +36,7 @@ public interface CurveBackend {
 	
 	MinPlus getMinPlus();
 	
-	CurvePwAffine getCurve();
+	CurvePwAffine getCurveFactory();
 	
 	LinearSegment createLinearSegment(Num x, Num y, Num grad, boolean leftopen);
