+++ src/main/java/de/uni_kl/cs/discodnc/CurveBackend_DNC.java
@@ -1,6 +1,5 @@
-package de.uni_kl.cs.discodnc.nc;
+package de.uni_kl.cs.discodnc;
 
-import de.uni_kl.cs.discodnc.CurveBackend;
 import de.uni_kl.cs.discodnc.curves.CurvePwAffine;
 import de.uni_kl.cs.discodnc.curves.LinearSegment;
 import de.uni_kl.cs.discodnc.curves.dnc.Curve_DNC;
@@ -9,7 +8,7 @@ import de.uni_kl.cs.discodnc.minplus.MinPlus;
 import de.uni_kl.cs.discodnc.minplus.MinPlus_DNC;
 import de.uni_kl.cs.discodnc.numbers.Num;
 
-public enum CurveImpl_DNC implements CurveBackend {
+public enum CurveBackend_DNC implements CurveBackend {
 		DNC;
 	
 		@Override
@@ -18,7 +17,7 @@ public enum CurveImpl_DNC implements CurveBackend {
 		}
 
 		@Override
-		public CurvePwAffine getCurve() {
+		public CurvePwAffine getCurveFactory() {
 			return Curve_DNC.getFactory();
 		}
